+++ examples/gemini_extractor_with_rag/src/main.rs
@@ -42,7 +42,7 @@ struct QuestionnaireResponses {
 
 struct QuestionnaireRag<I> {
     index: I,
-    samples: usize,
+    samples: u64,
 }
 
 impl<M, I> AgentHook<M> for QuestionnaireRag<I>
@@ -70,7 +70,7 @@ where
 
         let request = VectorSearchRequest::builder()
             .query(query)
-            .samples(self.samples as u64)
+            .samples(self.samples)
             .build();
         match self.index.top_n(request).await {
             Ok(results) => CompletionCallAction::patch(RequestPatch::new().extra_context(
