+++ examples/rag/src/main.rs
@@ -25,7 +25,7 @@ struct WordDefinition {
 
 struct DictionaryRag<I> {
     index: I,
-    samples: usize,
+    samples: u64,
 }
 
 impl<M, I> AgentHook<M> for DictionaryRag<I>
@@ -53,7 +53,7 @@ where
 
         let request = VectorSearchRequest::builder()
             .query(query)
-            .samples(self.samples as u64)
+            .samples(self.samples)
             .build();
         match self.index.top_n(request).await {
             Ok(results) => CompletionCallAction::patch(RequestPatch::new().extra_context(
