+++ tests/integrations/lancedb/mod.rs
@@ -27,7 +27,7 @@ mod fixture;
 
 struct LanceDbRag<I> {
     index: I,
-    samples: usize,
+    samples: u64,
 }
 
 impl<M, I> AgentHook<M> for LanceDbRag<I>
@@ -55,7 +55,7 @@ where
 
         let request = VectorSearchRequest::builder()
             .query(query)
-            .samples(self.samples as u64)
+            .samples(self.samples)
             .build();
         match rig::vector_store::VectorStoreIndexDyn::top_n(&self.index, request).await {
             Ok(results) => CompletionCallAction::patch(RequestPatch::new().extra_context(
@@ -448,7 +448,7 @@ async fn agent_with_application_defined_context_test() {
         .into_agent_builder()
         .add_hook(LanceDbRag {
             index: vector_store_index,
-            samples: top_k,
+            samples: top_k as u64,
         })
         .build();
